refactor: modularize codebase by extracting type definitions into dedicated files across all service modules
