+++ apps/gateway/src/modules/admin/admin.types.ts
@@ -1,0 +1,6 @@
+export interface GetAgentsOptions {
+  page: number;
+  limit: number;
+  status?: string;
+  search?: string;
+}
