+++ apps/gateway/src/modules/admin/index.ts
@@ -1,4 +1,5 @@
 import adminRouter from "./admin.routes";
 
 export { adminRouter };
+export * from "./admin.types";
 export default adminRouter;
