+++ apps/gateway/src/modules/agent/agent.types.ts
@@ -1,0 +1,4 @@
+export interface UpdateAgentProfileInput {
+  name?: string;
+  phoneNumber?: string;
+}
