+++ apps/gateway/src/modules/agent/index.ts
@@ -1,4 +1,5 @@
 import agentRouter from "./agent.routes";
 
 export { agentRouter };
+export * from "./agent.types";
 export default agentRouter;
