+++ apps/gateway/src/modules/analytics/analytics.types.ts
@@ -1,0 +1,9 @@
+export interface ConversationTrendRecord {
+  createdAt: Date;
+  updatedAt: Date;
+  closedAt?: Date | null;
+  status: "open" | "pending" | "resolved" | "closed";
+  metadata?: {
+    statusUpdatedAt?: Date | string | null;
+  };
+}
