+++ apps/gateway/src/modules/auth/auth.service.ts
@@ -11,6 +11,7 @@ import { generateOTP, hashOTP, verifyOTP as checkOTP } from "@shared/security/au
 import crypto from "crypto";
 import mongoose from "mongoose";
 import { OrganizationService } from "../organization/organization.service";
+import { SignupContext } from "./auth.types";
 
 export class AuthService {
 
@@ -48,15 +49,7 @@ export class AuthService {
     organizationName: string;
     password: string;
   }) {
-    type SignupContext = {
-      userId: string;
-      userName: string;
-      userEmail: string;
-      organizationId: string;
-      organizationName: string;
-      organizationSlug: string;
-      organizationPlan: string;
-    };
+
 
     const normalizedEmail = data.email.toLowerCase();
     let signupContext: SignupContext | undefined;
