+++ apps/gateway/src/modules/auth/auth.types.ts
@@ -1,0 +1,9 @@
+export interface SignupContext {
+  userId: string;
+  userName: string;
+  userEmail: string;
+  organizationId: string;
+  organizationName: string;
+  organizationSlug: string;
+  organizationPlan: string;
+}
