+++ apps/gateway/src/modules/contacts/contacts.service.ts
@@ -1,27 +1,6 @@
 import { Types } from "mongoose";
 import { Contact, Conversation } from "@shared/models";
-
-interface ListContactsOptions {
-  search?: string;
-  limit?: number;
-}
-
-interface UpsertFromAIInput {
-  organizationId: string;
-  conversationId: string;
-  name?: string;
-  email?: string;
-  phone?: string;
-  company?: string;
-  tags?: string[];
-  note?: string;
-  status?: "active" | "inactive" | "blocked";
-  sentiment?: "positive" | "neutral" | "negative";
-  summary?: string;
-  topics?: string[];
-  timelineLabel?: string;
-  timelineDetail?: string;
-}
+import { ListContactsOptions, UpsertFromAIInput } from "./contacts.types";
 
 export class ContactsService {
   async listContacts(organizationId: string, options: ListContactsOptions = {}) {
