+++ apps/gateway/src/modules/conversation/conversation.types.ts
@@ -1,0 +1,17 @@
+export interface ListConversationsOptions {
+    status?: string;
+    limit?: number;
+    offset?: number;
+    assignedTo?: string | null;
+}
+
+export interface UpdateVisitorInfoInput {
+    name?: string;
+    email?: string;
+    sessionId?: string;
+}
+
+export interface RouteConversationInput {
+    agentId?: string;
+    reason?: string;
+}
