+++ apps/gateway/src/modules/conversation/index.ts
@@ -1,4 +1,5 @@
 import conversationRouter from "./conversation.routes";
 
 export { conversationRouter };
+export * from "./conversation.types";
 export default conversationRouter;
