+++ apps/gateway/src/modules/membership/index.ts
@@ -1,3 +1,4 @@
 export { membershipRouter } from "./membership.routes";
 export { MembershipController } from "./membership.controller";
 export { MembershipService } from "./membership.service";
+export * from "./membership.types";
