+++ apps/gateway/src/modules/membership/membership.types.ts
@@ -1,0 +1,8 @@
+import { MembershipRole } from "@shared/models";
+
+export interface InviteMemberInput {
+    email: string;
+    name: string;
+    role: MembershipRole;
+    password?: string;
+}
