+++ apps/gateway/src/modules/organization/index.ts
@@ -3,3 +3,4 @@ export { OrganizationController } from "./organization.controller";
 export { OrganizationBillingController } from "./organization.billing.controller";
 export { OrganizationService } from "./organization.service";
 export { organizationSchema } from "./organization.schema";
+export * from "./organization.types";
