+++ apps/gateway/src/modules/organization/organization.types.ts
@@ -1,0 +1,11 @@
+export interface CreateOrganizationInput {
+    name: string;
+    slug?: string;
+}
+
+export interface UpdateOrganizationInput {
+    name?: string;
+    slug?: string;
+    logoUrl?: string;
+    whiteLabelEnabled?: boolean;
+}
