+++ apps/gateway/src/modules/storage/storage.service.ts
@@ -12,26 +12,11 @@ import {
   listObjects,
 } from "@shared/utils/storage";
 import type { Readable } from "stream";
-
-export interface PresignedUrlResponse {
-  uploadUrl: string;
-  downloadUrl?: string;
-  fileKey: string;
-  fileName: string;
-  expiresIn: number;
-}
-
-export interface FileMetadata {
-  fileName: string;
-  fileSize: number;
-  mimeType: string;
-  fileKey?: string;
-}
-
-export interface ProxyFilePayload {
-  contentType: string;
-  stream: Readable;
-}
+import {
+  PresignedUrlResponse,
+  FileMetadata,
+  ProxyFilePayload,
+} from "./storage.types";
 
 class StorageService {
   // ── Public URL ─────────────────────────────────────────────────────────────
