+++ apps/gateway/src/modules/widget/widget.types.ts
@@ -1,0 +1,3 @@
+export type ServiceError = Error & { statusCode?: number };
+
+export type AIInteractionSource = "widget" | "qr" | "link";
